+++ NineLanCacheUI/src/app/(pages)/RecentSteamDownloads/page.tsx
@@ -13,7 +13,7 @@ import {
 } from "@/components/ui/table";
 import { Input } from "@/components/ui/input";
 import { formatBytes } from "@/lib/Utilities";
-import { getSignalRConnection, startConnection } from "../../lib/SignalR";
+import { getSignalRConnection, startConnection } from "../../../lib/SignalR";
 import { imageCache } from "@/lib/ImageCache";
 import { AnimatedPage, AnimatedCard } from "@/components/animations";
 import { DateRangeFilter } from "@/components/DateRangeFilter";
@@ -115,7 +115,6 @@ const PreloadableImage = ({ appId }: { appId: number }) => {
             onError={handleError}
             onLoad={handleLoad}
             quality={75}
-            priority={false}
           />
         )}
       </a>
@@ -123,7 +122,7 @@ const PreloadableImage = ({ appId }: { appId: number }) => {
   );
 };
 
-export default function RecentDownloads() {
+export default function RecentSteamDownloads() {
   const [data, setData] = useState<DownloadEvent[]>([]);
   const [selectedRange, setSelectedRange] = useState(
     () => getStoredFilters()?.selectedRange || "0",
@@ -144,7 +143,6 @@ export default function RecentDownloads() {
     if (!filterText) return true;
     const searchLower = filterText.toLowerCase();
     return (
-      item.cacheIdentifier.toLowerCase().includes(searchLower) ||
       item.clientIp.toLowerCase().includes(searchLower) ||
       (item.steamDepot?.steamApp?.name || "").toLowerCase().includes(searchLower) ||
       (item.downloadIdentifier?.toString() || "").includes(searchLower)
@@ -179,7 +177,9 @@ export default function RecentDownloads() {
       params.append("excludeIPs", excludeIPs.toString());
       params.append("limit", "100");
 
-      const res = await fetch(`/api/proxy/RecentDownloads/GetRecentDownloads?${params.toString()}`);
+      const res = await fetch(
+        `/api/proxy/RecentDownloads/GetRecentSteamDownloads?${params.toString()}`,
+      );
       if (!res.ok) throw new Error("Failed to fetch data");
 
       const newData: DownloadEvent[] = await res.json();
@@ -205,7 +205,9 @@ export default function RecentDownloads() {
       params.append("excludeIPs", excludeIPs.toString());
       params.append("limit", "20");
 
-      const res = await fetch(`/api/proxy/RecentDownloads/GetRecentDownloads?${params.toString()}`);
+      const res = await fetch(
+        `/api/proxy/RecentDownloads/GetRecentSteamDownloads?${params.toString()}`,
+      );
       if (!res.ok) throw new Error("Failed to fetch new data");
 
       const newData: DownloadEvent[] = await res.json();
@@ -231,12 +233,16 @@ export default function RecentDownloads() {
 
   useEffect(() => {
     const connection = getSignalRConnection();
-    connection.on("UpdateDownloadEvents", fetchAndMergeNewData);
 
+    const handler = () => {
+      fetchAndMergeNewData();
+    };
+
+    connection.on("UpdateDownloadEvents", handler);
     startConnection();
 
     return () => {
-      connection.off("UpdateDownloadEvents", fetchAndMergeNewData);
+      connection.off("UpdateDownloadEvents", handler);
     };
   }, [fetchAndMergeNewData]);
 
@@ -246,7 +252,9 @@ export default function RecentDownloads() {
         className="p-4 mx-auto rounded-3xl bg-background"
         style={{ width: "98%", maxWidth: "1600px" }}
       >
-        <h1 className="text-3xl font-bold mb-4 text-center text-foreground">Recent Downloads</h1>
+        <h1 className="text-3xl font-bold mb-4 text-center text-foreground">
+          Recent Steam Downloads
+        </h1>
 
         <DateRangeFilter
           selectedRange={selectedRange}
@@ -262,7 +270,7 @@ export default function RecentDownloads() {
           <Card className="bg-card border-border">
             <CardHeader>
               <div className="flex items-center justify-between">
-                <CardTitle className="text-foreground">Recent Downloads</CardTitle>
+                <CardTitle className="text-foreground">Recent Steam Downloads</CardTitle>
                 <Input
                   placeholder="Search..."
                   value={filterText}
@@ -276,13 +284,6 @@ export default function RecentDownloads() {
                 <Table>
                   <TableHeader>
                     <TableRow className="border-border hover:bg-transparent">
-                      <TableHead
-                        className="text-foreground cursor-pointer hover:text-muted-foreground text-center"
-                        onClick={() => handleSort("cacheIdentifier")}
-                      >
-                        Service{" "}
-                        {sortField === "cacheIdentifier" && (sortDirection === "asc" ? "↑" : "↓")}
-                      </TableHead>
                       <TableHead
                         className="text-foreground cursor-pointer hover:text-muted-foreground text-center"
                         onClick={() => handleSort("createdAt")}
@@ -301,6 +302,9 @@ export default function RecentDownloads() {
                       </TableHead>
                       <TableHead className="text-foreground text-center">Hit %</TableHead>
                       <TableHead className="text-foreground text-center">Miss %</TableHead>
+                      <TableHead className="text-foreground text-center">
+                        Download Progress
+                      </TableHead>
                     </TableRow>
                   </TableHeader>
                   <TableBody>
@@ -313,10 +317,13 @@ export default function RecentDownloads() {
                       const hitPercent = total > 0 ? (row.cacheHitBytes / total) * 100 : 0;
                       const missPercent = total > 0 ? (row.cacheMissBytes / total) * 100 : 0;
                       const appId = row.steamDepot?.steamAppId;
+                      const totalBytes = row.totalBytes ?? 0;
+                      const downloaded = (row.cacheMissBytes ?? 0) + (row.cacheHitBytes ?? 0);
+                      const downloadPercent =
+                        totalBytes > 0 ? Math.min((downloaded / totalBytes) * 100, 100) : 0;
 
                       return (
                         <TableRow key={row.id} className="border-border hover:bg-secondary">
-                          <TableCell className="text-foreground">{row.cacheIdentifier}</TableCell>
                           <TableCell className="text-foreground text-sm whitespace-nowrap">
                             <div>
                               {formatDateTime(created)} → {formatDateTime(updated)}
@@ -370,6 +377,29 @@ export default function RecentDownloads() {
                               <div className="text-xs mt-1 text-foreground text-center">
                                 {formatBytes(row.cacheMissBytes)} • {missPercent.toFixed(1)}%
                               </div>
+                            </div>
+                          </TableCell>
+                          <TableCell>
+                            <div className="w-full min-w-[150px]">
+                              {totalBytes === 0 ? (
+                                <span className="text-red-400 text-sm">
+                                  Could not find Steam Manifest
+                                </span>
+                              ) : (
+                                <>
+                                  <div className="h-4 bg-muted rounded overflow-hidden">
+                                    <div
+                                      className="h-full bg-indigo-500"
+                                      style={{ width: `${downloadPercent}%` }}
+                                    ></div>
+                                  </div>
+                                  <div className="text-xs mt-1 text-foreground text-center">
+                                    {downloadPercent < 100
+                                      ? `${formatBytes(downloaded)} / ${formatBytes(totalBytes)} (${downloadPercent.toFixed(1)}%)`
+                                      : `${formatBytes(downloaded)} (100%)`}
+                                  </div>
+                                </>
+                              )}
                             </div>
                           </TableCell>
                         </TableRow>
